+++ tests/testthat/test_dir_recursive/test-bare-expectations.R
@@ -1,0 +1,1 @@
+expect_equal(2, 2)